move #add_coin to CoinSorter module
