+++ models/coin_sorter.rb
@@ -10,4 +10,16 @@ module CoinSorter
       return nil
     end
   end
+
+  def add_coin(coin)
+    coin_type = CoinSorter.valid_type(coin)
+
+    if coin_type
+      @coins[coin_type] += 1
+      @inserted_value += coin_type
+      show_inserted_value
+    else
+      return false
+    end
+  end
 end
