+++ spec/machine_spec.rb
@@ -6,7 +6,6 @@ describe Machine do
   let(:quarter) { Coin.new({weight: 5.67, diameter: 24.26, thickness: 1.75})}
   let(:dime) { Coin.new({weight: 2.27, diameter: 17.91, thickness: 1.35})}
   let(:nickel) { Coin.new({weight: 5, diameter: 21.21, thickness: 1.95})}
-  let(:penny) { Coin.new() }
   let(:chips) { Product.new({name: "Chips", price: 50}) }
 
   describe "#initialize" do
@@ -16,22 +15,6 @@ describe Machine do
 
     it "initializes with no stored coins" do
       expect(machine.coins).to eq({25 => 0, 10 => 0, 5 => 0})
-    end
-  end
-
-  describe "#add_coin" do
-    it "stores the coin in the machine" do
-      machine.add_coin(quarter)
-      expect(machine.coins[25]).to eq 1
-    end
-
-    it "adds the coin's value to the inserted value" do
-      machine.add_coin(dime)
-      expect(machine.inserted_value).to eq 10
-    end
-
-    it "does not add invalid coins" do
-      expect(machine.add_coin(penny)).to be false
     end
   end
